requirements.txt
Flask
flask-cors
httpx
beautifulsoup4
lxml
gspread
oauth2client
python-dotenv



